requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
Pillow
psycopg2
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions
uritemplate==4.1.1
gunicorn
whitenoise
cloudinary
django-cloudinary-storage
django-import-export
python-decouple
social-auth-app-django
django-rest-framework-social-oauth2
djangorestframework-gis
